replaced Bytes method with Export
